requirements/dev.txt
-r prod.txt
-r test.txt

django-debug-toolbar